(mingle-fr-2012-66) Email sent to MaxMind should be an MD5
Email sent to MaxMind should be an MD5 of the email address, and not the actual email address.

https://mingle.corp.wikimedia.org/projects/fundraiser_2012/cards/66

Email addresses are now being sent as an MD5.

All required and optional fields are being properly used.

Removed unused class member: $this->log_msg_prefix

It was copied from Gateway_Extras_CustomFilters and was causing failures.

Change-Id: I6cd8bdd2c3289e426eefd904fd659a82acf2bd1b