(mingle-fr-2012-67) Upgrade our implementation of MaxMind minFraud
The minFraud libraries were upgraded to v1.50.

There were not that many changes to this library from MaxMind.

There were a few deprecated methods replaced. The rest of the changes restore the whitespace that was converted to tabs or removes whitespace that was added for readabililty. The goal is to make the libraries as similar as possible to the original.

The primary change to the code is the elimination of score.

Change-Id: If6da06cefad8211aa81139aefb086b0b1c3e25c5