Amazon Simple Payment integration, FR #43
Also, small changes to the GatewayAdapter base class:
* Declared some more pure abstract functions on the interface,
because they are.
* Refactor getFormattedResponse into the base class because it's
generic and I wanted to reuse it.  Marked overridden definitions
as deprecated.
* Some globalcollect-specific strings are removed or marked.

DonationData
* Removed default payment_type of 'CC'. This must now always be
in session data. Tested with globalcollect and it seems to work.

Patchset 5:
* remove rapidhtml stuff again
* remove getFormattedResponse where it's redundant
* don't try to displayForm cos there is none

Patchset 6:
* add example global config variables
* rename stupid redirect param

Patchset 7: rebase

Patchset 8: code style + two string typos

SNAFU -- error behavior is probably incorrect when a) the special
page is entered without PaypalRedirect=1, or b) the response from
Amazon is failure or has an incorrect signature.  I think we will
want to redirect to an error page in these cases.

Change-Id: I5e481d745de1c94e13ed33755af74c74aca94d0a