(FR# 1084) Whitelist for Paypal Xclick
Paypal in China is unable to offer ChinaUnionPay with the
cmd=_donations workflow. Therefore we need to have a list
of countries that must use the cmd=_xclick workflow and
serve that for those countries.

This is now possible with the $wgPaypalGatewayXclickCountries
parameter.

Change-Id: I8e98b794f7c7f386572d3b47a6065051289e40bc
(cherry picked from commit 160eb8228315ad07fb01fc4787eac7816746fda9)