Dyanmic Error Pages, Part 1
Shouldn't functionally change anything we have at the moment, so it
should be safe to merge.
The first step toward having dynamic error pages. This also fixes the
back link so it will go to the most recently used payment form, instead
of some crazy place on donate.
What's missing: The switch that will enable us to use this error page,
instead of the static one.
Also, the appeal param isn't sticky, so if you're relying on css in
on-wiki appeals to alter forms, STOP IT.

Change-Id: I6b1baffd664ae814740bf57774c9f62859a3eefc