Migrate amazon_gateway.i18n.php to JSON i18n
This was missed in 0e1f29e675393e9dcf because it is not actually used
anywhere in the extension. I'd expected it to be used in
DonationInterface.php in a block of code similar to this at line 1038:

 if ( $optionalParts['Adyen'] === true )

Procedure per https://www.mediawiki.org/wiki/Manual:GenerateJsonI18n.php
with shim.

Change-Id: I45ece1dbda5b4cdfcb4a833e13da983fd63473c3