Some Transaction / Transaction Reponse Cleanup
I was trawling through gateway adapter and found some stuff
that could be removed / done in a more canonical way.

Also fixed a PHP deprecated function in the GC adapter.

I was going to completely extract part of do_transaction_internal
into it's own method that could be overridden, but I decided instead
to scrap that.

Change-Id: I7a885693b15c21124479ca22e7118085f81fb050