Adding two new queues/messages.
These will be the first messages to be shoveled into the new
"fredge" database:  Basically mirroring the data we currently
stash in the payments-initial and payments-fraud syslog buckets.

PS2: Removed data that will be redundant once this data lands
in the db, and we can easily join with contribution_tracking.
Added user_ip to the antifraud message.
Removed naming inconsistencies.

Change-Id: I748e6b3db77a9663c0e4b90d9b8985d08709a162