Merge standalone framework
Wraps MediaWiki core functions in an abstraction layer "WmfFramework", so we
can access payment processor interfaces from Drupal...

Presentation code was not included in this port, that might come in
a later commit.

Should have no effect on behavior of the MediaWiki extension.

DEPLOYMENT: the recurring_globalcollect module needs patching before using this
library
* We should create the adapter in batch mode, and pass donation data using
a different parameter than in the standalone library.
* Populate contribution_tracking_id to avoid DI ContributionTracking dependencies.
* Pass some more bluff parameters, to circumvent unported, MW-only code paths:
	'language' => 'en',
	'returnTo' => null,
	'returnTitle' => null,

Change-Id: I46e64a5a9548e4edbb02712332bd9a8359f142f8