DonationData::sanitizeValue: drop extra args
HHVM currently has a bug whereby if a callback function passed to array_walk
has an arity greater than two, the names of arguments past the second argument
aren't bound. It's a bug in HHVM rather than DonationData. However, you're only
calling this function once, and you're not passing it the extra args, so it
costs us nothing, really, to hard-code them in.

Also change a bare '$matches;' declaration to actually declare the var by
assigning it null.

Change-Id: I0261979fde9bbdb6f3791b9decbd9928c8f7d796