Add a grunt linting checker for jshint, jscs, and banana and pass them
These are configured using the Wikimedia-standard presets for code style and sanity-
checking. The intent is, once this is passing, to add this as a (voting) job to the
gerrit merge system.

To test locally, checkout and run:

	npm install -g grunt-cli
	grunt test

Follow-up work should add csslint and various others, and work through some of the
hacks required to get this passing.

Note: The "payflowpro_gateway" directory has been excluded from this work because it
is apparently unused in production.

Change-Id: If26d95fc9f4141f558e54cbb4ce5174c7cb111da