Cleaning up the payments error logs
delete tests for deploy

We really don't need to see an entry in the payments error log
for every invalid or fat-fingered credit card payment attempt.
Mostly we should reserve error messages for things that we don't
expect to happen all the time, or that might indicate a programmer
error.

Change-Id: I1eb2f4388b94e49b2734107609e21c7f40aed43d