Split staging from unstaging
The combined stage_ helpers were muddying things, and were often only
working correctly when used in the unstaged -> staged direction.

Many of these functions were working on staged source data, when they
should have been working on the unstaged data.  This was mostly working
by accident, because we copy the normalized data on top of both the staged
and unstaged data before the staging code runs.

Change-Id: If2d241fed6cc055123abfc97d58e25bdd34ae0f8