"language" parameter will work without "uselang"
We have a sketchy legacy "language" parameter, which is interpreted by
DonationInterface but not by MediaWiki core.  Core only respects "uselang".
There are some code paths which accidentally don't include the "uselang"
parameter, so this patch causes "language" to have the same effect on
global UI language as a backwards-compatibility workaround.

Bug: T94604
Change-Id: I15ed1b5656c7679eb379783d331bd326683a78a2