Prepare to prune $transaction_results
Free-form array was growing like kudzu - replace it with an object
called $transaction_response, to clarify that it holds a processed
response, not the final result of the transaction
Get rid of a couple of random top-level keys.
Fix and deprecate getter/setter shims.
We want processResponse to eventually create this whole thing.
Replacing array_key_exists with isset where it matters downstream.
TODO: get rid of less-used properties

Bug: T98063
Change-Id: I0191861e20a537534d2a96eb1ca8730df0033307