Use ct_id.numAttempt format for Astropay order number
We don't get a gateway txn ID till the payment goes through, so
we need this number to correlate donor details with payment listener
messages or audit lines.  Worldpay adapter had code that does just
that, so move it into parent class and make it a order_id_meta option.

Change-Id: Ie5374b37564e7b1c0b6b6db8a4046d407e7ea5b5