Parse more of AstroPay's error descriptions
AstroPay's error codes are still unreliable and not informative enough
to drive our decisions.  But at least we have a list of error
descriptions with suggested ways to proceed.

This patch deals with invalid CPF errors, over-the-limit errors,
blacklisted user errors, and 'could not register' (capacity?) errors.

Also, log the whole response when AstroPay returns an error, and tweak
css so long error messages at the top don't widen the form.

Bug: T106053
Change-Id: I97db8f6216d75a0da508af713ef02616d8fe90d1