Fix fail page redirection (see deploy note!)
DEPLOYMENT NOTE: changed $wgDonationInterfaceRapidFailPage to
$wgDonationInterfaceRapidFail

Production is currently sending users to /Donate-error, which is
a 404.  Might have been a side effect of the 1.25 upgrade.

Add default value and doc string for RapidFail global.
Consistently use GatewayPage->displayFailPage().
Use Title::newFromText and ->getFullURL() to make the correct URL,
always append uselang parameter.  No need for the if ($page) check
since we set a fallback default in DonationInterface.php

TODO: create Special:DonationFailed so as not to depend on wiki
content.
FIXME: production fail page Donate-error is not translated at /lang
like production thank you pages, so we are appending language
inconsistently

Bug: T108041
Change-Id: I9784580cdc9d15689267c44db9d98492565840c9