Blank submethod on method change, OID on recur change
Sharing order IDs between recurring and non-recurring leads to
problems with GlobalCollect, and a submethod from session that
doesn't match the method on a new request leads to invalid ffname
log spam.

This is the quick and dirty version of the fix, duplicating some of
DonationData's normalization code in the check/reset function.
A better fix will involve decoupling DonationData from adapter so
that its constructor doesn't have side effects like finalizing the
adapter's order ID.

Bug: T105041
Bug: T108248
Change-Id: Ie58559a9c7e657312432d4141fa72b0f1691ad69