Make it easier to find donations by email in AstroPay
We send them a truncated email address as customer id, 'cause that's
what they let us search on in the merchant console.  If a donor
includes a spam-tracking subaddress, we should honor that in our
communications, but when donor services gets an email from them and
wants to find their donation in the console, they probably won't
get the subaddress.

Change-Id: If92a0d5843855e2054c8af89723bef4cfea4f057