Limit OID reset on recurring changes, log resets
We were resetting OID for recurring donors at the resultswitcher
because there was no recurring parameter on the querystring.  This
kicked them out since the OID wasn't in session, and many of them
made duplicate donations.

Only reset if there is a recurring parameter on the query string.
The form chooser will add an empty recurring parameter for new
one-time donation attempts.

Log changed values every time we trigger a reset.

Bug: T110367
Change-Id: Iffa96ebf1a328820cfd3bb8ee0f3fea9780f52cb