Don't remove overlay when redirecting
After the API call comes back, only remove the overlay if we're
going to let the user try again.  If we remove it before redirecting
the user to the thank you or fail page, it looks like the click
did nothing for however long it takes to load the next page.

Change-Id: I0f561f5c2674d87eb9503eb7b9ca9b38364b4e0e