Amazon: check CSRF token, preserve data after expiry
Ensure api.php posts have a valid token, and reload the page with
all the hidden fields on the URL if it expires.  Reload the same
way when there's an issue with the Amazon widgets.

Re-using the expired token message from the Worldpay CC token patch,
which is worded for an alert box.

Bug: T117827
Change-Id: I6c4f95348067b9af203b81b86e2b8c240b9be69a