Replace rubbish data in currency code
Can't just blank out everything that fails validation, because
a lot of it could be almost correct.  With this logic, we make
sure that we have a valid ISO code after normalization.
Then if it fails validation, we know that means the adapter
doesn't support it, and we can attempt to convert to a fallback
currency.

Hat tip to cdentinger for investigating the test fails.

Bug: T119120
Change-Id: I1cc2d111c83d19b978e59caa5b64be1ba5405f4b