Remove whitebox wiki template from forms
Remove the whitebox wiki template [1] from forms where it is still used. This
template included info on "Where your donation goes" but is very out of date,
especially in its translated versions (e.g. in referring to Wikipedia as
5th most visited site on the Internet, and staff numbers). It is already not
used on our main credit card forms. Discussing with creative, we decided to
remove it from the remaining forms too.

[1] https://payments.wikimedia.org/index.php/Template:2011FR/core-appeal-whitebox-nodiv/en

Change-Id: I9692f8d836be42fae39824f9a9083718c74836d7