Handle another GC error indicating Order ID reuse
We got confirmation that error 400120 happens when we try to reuse an Order ID.
Regenerate the Order ID when we see this error.

TODO:
* I didn't manage to get this test to fail with the old logic, which makes me sad.
For some reason, we responded incorrectly to the NOK because it didn't throw an
exception.  I don't want to fix broken code to make it break better, so... moving
right along.

Bug: T120805
Change-Id: I736b4a99e6ca8bbd117419081a81359ce3379209