Fix 'emailAdd' / 'email' confusion
'email' is the authoritative field name

Gets rid of some special cases and confusion, including a bad
validation message for blank emails. For everything else,
donate_interface-error-msg-X is the label for field X, but
donate_interface-error-msg-email was the full explanation that
you used an invalid format, while donate_interface-error-msg-emailAdd
was the label.

Duplicated key donate_interface-donor-emailAdd was not translated
in any language that lacked donate_interface-donor-email.

Still allows emailAdd as an input, though that could probably go too.

Change-Id: I2c44e3576964be235f0503cc313fef6ecca64050