More cleanup around GatewayPage constructor
Set the language up front, so we have it for the displayFailPage in
the catch block. Remove isFailed flag and just do the stuff in catch.

Also move the hook registration down to where we know we're going to
try showing an actual adapter form.

Finally, clean up some test constructor stuff we don't need since
awight fixed the page titles.

Change-Id: I581ebeae5c2b6a1726082b9bb37cdcecd91463c8