Less overloaded processResponse
This gets us closer to upholding the single responsibility principle.

processResponse should be for doing stuff with responses to API calls

New function processDonorReturn is for doing stuff when the donor
hits the resultSwitcher

This will clear the way for us to make Ingenico's result switcher
use the common GatewayPage::handleResultRequest code and DRY up the
framebusting in a way that helps prevent unintended duplicate donations.

Bug: 153972
Change-Id: I275b91068374387bd623c10c8234a3d122ae4344