Remove unused api_request stuff
This was used to skip a bunch of RapidHTML logic, which is now gone.

We were also consulting this in the Amazon adapter to decide if we
had to push a ffname into the session. I believe this is now
unneeded as well. If I'm wrong, the worst that should happen is we
get spam in the error logs on the Amazon form when we return from
their login page.

Bug: T87043
Change-Id: I2eb79b9ad3afd9983c52c5aa271edc4095d6809f