Deprecate "manual" errors and shift responsibility for validation into the adapter
The original intent may have been a firewall between the validation errors
which can only be set by the adapter, and errors which can be set by other
classes.

I'm simplifying down to a single list of errors.  If any errors are present,
you will not advance to the next step in your donation.

GatewayAdapter takes responsibility for running validations, away from
DonationData and GatewayPage.

Bug: T160385
Change-Id: I279052521a20e28d0935ce100c31087a2519e365