Move fraud filter config to TestConfiguration
Once upon a time I thought it was OK to do stuff in test case
constructors. That was totally wrong, as all of the constructors
are called before any of the tests are run. These maps don't need to
be set dynamically - they can just live in the test config.

Change-Id: I33c1260a8d7394ca1722d70eff0bf2c8ddae7a9f