PayPal EC: use ajax to get redirect URL
Act like Adyen or Ingenico, showing the form and using and AJAX call to do
the server-to-server setup in the background and return a formaction URL.

Slightly different from Adyen and Ingenico, this will submit instantly and
redirect instead of showing an iframe.

This might solve our dead session errors.

Bug: T167923
Change-Id: Ifad296dbd09dc6935ecd8799278e411ef3b14548