PayPal EC recurring: initial charge is part of subscription
We were originally doing a kludgey workaround that had an initial
standalone payment, then a subscription starting in a month. That
made it really hard for Donor Services to find subscriptions.

PayPal techs told us the reason for the kludge is gone for 99% of
users, and we're waiting on a list of countries where it still
turns up.

Bug: T170478
Change-Id: Ifebc76e0e501303381ff8c73000d02e36cdb796c