PayPal EC recurring: initial donation is async
After we set up a subscription, the initial donation isn't immediately
charged. So we don't have a donation to send to the queue here, just
a subscr_id. We're stashing this in the pending queue now, for potential
use later.

Note that this is actually how all donations work with the legacy PayPal
integration - the payments cluster just sets things up and lets the IPN
listener handle the payment notifications and getting stuff into Civi.

Bug: T170478
Change-Id: I5f80586485d6a0cef53c5d19ce8913e61c258121