Order sequence numbers start at 1, not 0
It's way too easy for order 1234.0 to get rounded off to 1234 in
transit.

TODO: simplify the sequenceNumber stuff. We initially added that
for WorldPay, which required a unique ID for each API call, not
just each merchant. For everyone else, we can maybe just use the
attempt number (numAttempt). This patch attempts to preserve the
existing behavior re: sequence number.

Bug: T171891
Change-Id: I2ae477f3d1e4bb9d155333534b9fcb5c6a14074d