Fix double continue button situations
If there are errors in the personal info fields and the user tries
to select a submethod, do nothing other than highlight the errors.

If the user has a submethod selected and reloads the page, clear the
selection and don't show the continue button.

This removes the confusing situations where there could be two
continue buttons on screen (from this, and from the iframe).

Bug: T149366
Change-Id: If7405408659f588e506179781ed7ab5874c0a04a