Fix counting uncountables
This is necessary to make most of the DI tests run on php 7.2, which
throws an error when we count() something that's null or a string.

One instance was simple to replace the !count() with empty, and in the
other, the result is preserved by adding 1 for a scalar - in php 7.0
and below, count( "ALL" ) returns 1.

Change-Id: I86730722095bcc2ccde69dc4e8f0b93806f8a830