Add opt_in field for selected countries
DEPLOY NOTE: don't put this up in the front end till there's
something up on the back end to use it.

Shows the radio buttons for GB and IE for now. Should send an opt_in=1
or opt_in=0 in pending and donation messages when the choice is shown.
When the choice isn't shown, it shouldn't send any opt_in field at all.

Bug: T199278
Change-Id: I1711fe4cdfac0d6582e78fbf4b7c438814f4be4c