JS module to add English email explainer
This is for an English-only test to include the opt-in fields in the
banner. Should be deployed as a variant. To deploy as a variant, copy
country_fields.yaml and ui_modules.yaml from globalcollect's config to
a new variant directory, take opt_in out of the country_fields, and
add ext.donationInterface.emailExplain to ui_modules.

Bug: T200218
Change-Id: I8510617008c25b3eb8c5cf4494da98a1ec727f32