Show opt-in interstitial for paypal
Would be nice to have logic to skip this when there's a value on
the querystring.

Also moves the opt_in block to it's own mustache sub-template, and
hides the 'billing information' header if that's the only required
field.

TODO: refactor validation to let us run the rules validation without
showing the errors, to see if the rules are already satisfied on page
load.

Bug: T199281
Change-Id: Ib3e75bc8cc626a48c3f6d6178d398d1f7f7a515c