Override logo on certain conditions
Yet another way to mess with the display. This part is all in
the mediawiki-specific layer, so at least there's no need to
mess with the cross-platform abstractions. Since it's not affecting
the actual payment forms, and since we want the actual paths in the
settings, it felt wrong to do as a 'variant'.

Bug: T203603
Change-Id: I95aa7009f7a9cea9a6408260243754d38978b42b