Fix staging for donor locale
We weren't splitting out modifiers before tacking on the country code.
That meant that after staging language=zh-ha and country=US, we'd have
staged['language']='zh-ha_US'. In the Ingenico adapter, that would
then get trimmed down to five chars to fit, so back to 'zh-ha'.

New Ingenico API throws errors on this.

Bug: T206093
Change-Id: I6da278c78abd0648b3be1e120ff7b0907f11a835