Form chooser: fix zero-weighting logic
This logic was added so we could make forms available for testing
without making them preferred over other gateways with fewer card
options. It's currently a bit too aggressive, firing even when we
specify the gateway directly in the URL and leading us to redo
the selection without the gateway constraint.

Bug: T209416
Change-Id: I297e14a10a01dc9e00f3dbdef49cfef63c1ebd10