Make email explanation translatable
We started with this only in English-speaking countries, but we'll
need to use it in other languages soon.

To test, go to an ingenico, adyen, or globalcollect URL with
variant=emailExplain on the URL, e.g.

/index.php?title=Special:IngenicoGateway&payment_method=cc&currency=USD&amount=35&country=US&variant=emailExplain

A message should appear under the email field explaining that
we will send a donation receipt.

Bug: T215939
Change-Id: I27b2a2aab5f3871930d34d08bcf0dfaa9b783177