Hide opt_in buttons when given on querystring
When the opt_in value has been provided on the querystring, that means
the donor has already seen the form on donatewiki or in a banner (all
params from those sources are added to the querystring rather than POST).
In that case, we don't want to show it again on donatewiki.

When the opt_in value has been set by POST, that means we're re-displaying
the form after a submission, which might mean some data validation issue
in another field. In that case, we want to keep showing the form.

Bug: T217254
Change-Id: Ibbe36f1a9cdbb5291a13ac03b1f1b63256dd2eb1