Mark redirects 'liberated' so we process return immediately
For credit card donations, we assume the donor has entered their
data in an iframe, and that on return we will need to pop them out
of it to show them the TY page. The return pages set a 'liberated'
flag in session (as well as on the URL) when they load the
framebuster code.

This patch pre-emptively sets the 'liberated' session flag in the
API response whenver the response would make the Javascript send
the donor away with a full redirect. It also extends the Ingenico
redirect/iframe switch on 3D Secure to the legacy GlobalCollect
adapter.

Bug: T220512
Change-Id: Ic392eb3a1c74d742cdacc7ccd63061c287dc15e8