Rebuild email forms in DonationInterface
New forms are at Special:EmailPreferences/optin and
Special:EmailPreferences/unsubscribe. You can get the fancy version
of opt-in by going to Special:EmailPreferences/optin?v=wle_001.

Opt-in works, but not unsubscribe yet. That's actually OK, since
we'll have to leave the old unsubscribe up for a while to deal
with emails that are sitting in people's inboxes. We can just use
the new extension for opt-in till we've got the unsubscribe working,
at which point we can set up a redirect and get rid of the old
extension.

Bug: T218171
Change-Id: I96d8e2c64fa96b785fbf534937275e914f9d3d1e