Move requiredFields overrides to config
AstroPay (dLocal) forms aren't working with the optional patch,
because the adaptor was overriding getRequiredFields to add
the fiscal number depending on the country, and the template
rendering code was changed to call getFormFields instead.

This patch deletes Adyen and AstroPay's getRequiredFields overrides
and moves the fields that were added there to config, where the new
getFormFields method will pick them up.

Bug: T219558
Change-Id: Ie1f6950f95db7722522d0ec04287271731d7f3e1