Make 'emailExplain' default behavior
We can only apply one variant at once, so this frees us up to be
able to use the employer variant in countries where we would
otherwise need to use emailExplain.

With this patch, if you pass opt_in=0 on the query string (and
don't pass any variant), the email field should get an explanation
underneath it saying we only want it to send you a receipt.

Bug: T218730
Change-Id: I544838e54fb7ba67070604605c5408e53164a018