Allow overriding built-in yaml files without variant
This extends the 'variant'-style overrides to let you override the
defaults without passing a variant parameter or futzing with the
DonationInterface-repo code. It allows us to have some yaml changes
that only live on the production cluster.

Testing this is very similar to testing a variant: create a new
directory and point the LocalSettings.php global
$wgDonationInterfaceLocalConfigurationDirectory to it. Then in
that new directory, create a 'globalcollect' directory for example,
and copy the globalcollect country_fields.yaml into there. When you
modify your country_fields.yaml, those changes should override the
built-in country_fields.yaml (but in turn be overridden by any
country_fields.yaml in a loaded variant).

Bug: T225839
Change-Id: I02abfd5194419190db0e279a31ebfc033954afa1