RecurringConversion API
Initially creating this as a super-minimal API, just accepting
the gateway and amount, assuming that all the other parameters
will be the same. We shouldn't be allowing the donor to change
anything else at this point, and we rely on having the recurring
payment token in session in any case.

The 'gateway' param is used to instantiate the proper object,
but the amount is actually re-harvested from the POST in the
DonationData constructor, so we don't need to touch it in the
api execute method.

Bug: T216560
Change-Id: Ia056a8fd0c2331f1238f75a7bea447d0ab120112