Add form_variants dir to DonationInterface
DEPLOY NOTE: update $wgDonationInterfaceVariantConfigurationDirectory
in LocalSettings to point to this new directory. Then we can delete
the form_variants dir from the mediawiki repo.

Same logic applies as for why we moved the php FormSettings

We're the only users of this now, and it'll make testing things that
depend on variants much easier.

Change-Id: Ibe3f532ddcfb392dbd435c45d0f6a929e150c42f