Wire up 'yes' buttons for recurring conversion
To test this, you'll need to make an Ingenico Connect donation with
a variant starting with 'upsell'. Clicking the 'Yes' button should send
a subscr_start message to the recurring queue with the suggested amount,
while clicking the different amount button, filling in another amount,
and clicking the 'Donate Monthly' button should send a subscr_start
message to the recurring queue with that other amount

If the conversion is successful you should be redirected to the thank you
page with an extra parameter on the URL: recurringConversion=1.

If the conversion fails (i.e. no tokenized donation in session) you
should see a message saying your initial donation was successful but
that the recurring conversion failed. Dismissing that should send you
to the thank you page.

Bug: T216560
Change-Id: Id2e28084bba094f07bbad1d1ac49f2393978c731