Add div under updonate to disable form
Looks like this was already in the provided CSS, and I just forgot
to include the div in the initial commit.

The weird bit is that with a narrow window, the wikimedia logo /
home page link is still clickable, and shows through the modal window
( see https://phabricator.wikimedia.org/F30041467 ). I had to undo the
.mw-body-content's default position:relative (overriding it to 'inherit')
to fix it, following advice at https://coder-coder.com/z-index-isnt-working/
It doesn't seem to mess up the organization of the page (and only loads on
upsell), but I'd happily accept other suggestions.

Change-Id: I1315729b4cdf4ec3b1e774e849139123c78052d9