Form labels via new variant mechanism
This patch introduces another way to change the form via the 'variant'
parameter. Partial templates can now be included in the base dir of a
variant, and will replace the default partial templates.

Adds a variant to illustrate with labels atop contact information
fields. Amount and currency (when editable) aren't labeled to avoid
duplicating the 'Donation Amount' header, but could and maybe should be.

Kinda hokey to add all the empty files so the tests pass, but seems harmless.

Adds a max-width to the donate div so the new content in that div doesn't
widen it too much at the expense of the appeal.

Bug: T169795
Change-Id: Ibd7b5f842cb7060cf329ed43d1d6eb41d4f0b355