Cancel text for error form
Show a different message for donors who explicitly click cancel in
the processor frame (at least for Ingenico).

This is implemented via the error form definitions and corresponding
logic in the GatewayFormChooser.

We weren't using submethod-specificity (i.e. an error page just for
iDEAL or Amex), so I replaced it with payment-status-specificity and
clarified the logic a bit.

Adds getFinalStatus to the GatewayType interface alongside
getValidationAction. Also allows for a status other than 'failed'
(e.g. 'revised', 'cancelled') to emerge from a dead CC transaction
under Ingenico.

Bug: T224063
Change-Id: I0fd63e8f684cda86297413e818628f35810394ff