Forms CSS: less scroll, fix field widths, minor cleanup
Get rid of some rules referring to nonexistent elements and some
bottom padding. Hide the footer better. Eliminates the scroll bar
that appears without need when the links approach the bottom of the
page.

For halfwidth and thirdwidth form fields, share the space evenly
and don't leave an extra space hanging off the side. This was
previously a bit fuzzy because there were spaces in between
the divs, but we've tightened up the HTML to eliminate those, so
we can use exact measurements now.

Also combine a couple of rules and drop old -moz- prefix on
placeholder pseudo-element (not needed since FF 51).

Change-Id: I91cfb9d9b0bbbd6d548f80812ff05db062f2c4ac