Iframe liberator always reloads page
Deals with the following problem:
* Donor is using a processor which is sometimes processed in an iframe
and sometimes processed with a full redirect.
* Donor ends up using the full redirect. This should in theory store an
indicator in session saying the donor has already been popped out of
the iframe.
** SOMETHING BAD HAPPENS and the popped-out indicator is not set or the
donor's session dies
* On return, the indicator is not found and our server assumes the donor
needs to be popped out of an iframe, so only loads the liberator script
* Liberator script sees top.frames.length == 0 and so does nothing, not
even hide the UI elements
* Donor is stranded with a Log In link

This commit makes the liberator script ALWAYS reload the page with a
querystring indicator added. On the secord load, we will either process
the donation or deal with dead session / repeat processing.

Bug: T236803
Change-Id: I77ccdd70144e8133050d30b4910465479f59fc40