Monthly convert: better suggested amount, skip on min/max
Use suggested amounts for dollar and pound/euro from the banners,
and send the donor directly to the thank you page if they are
using an unsupported currency or an amount above the max or below
the min. To test this with the Ingenico test charges, you can use
the CAD / 1.40 test card to check the skip behaviour.

TODO: this shows a flash of the monthly convert ask without amount
before redirecting the donor to the TY page. Can we hide that?

Bug: T237775
Change-Id: I0e2cf8e268a8eb9e2ae4dc65af61d179d2e7c01a