Monthly convert: variant 008, different appeal paragraph text
Use variant=monthlyConvert_008 to test this. It should act just like the
normal monthlyConvert, but show the text 'Thank you for your gift. We
have to ask: are you able to do a bit more? Only a fraction of all our
readers make it to this point, because most don't donate. And of all our
donors, only a small portion consider an automatic monthly gift. But
monthly support is the best way to ensure that Wikipedia keeps thriving.
Will you help ensure our financial stability with a sustaining donation?'

Bug:T238144
Change-Id: Iec9d4d8f6191fcccf169bc3d0d3a70f9b7cb2c33