Ingenico: mark attempts processed by checkout ID
As a way to avoid double-processing the same donation in case the
donor hits the back button or reloads the result page, we save
a flag in session saying that a donation with a certain ID has been
processed. Since the hosted checkout ID is unique per attempt, and
always comes back on the query string, we can use that for Ingenico.

In practice, we're not actually double-charging people when they
reload, but we are making extra API calls that could be avoided.

Change-Id: I73ec0eeac0c73cd94002d013a160761b4e326a02